QCamera2: Force exit deferred allocation routine in error cases.
If there is any error while trying to start channel, deferred
buffer allocation would have been started but waiting for condition
signal. In working cases, this thread is signalled only after start
channel.

But as part of channel start failure, interface will eventually call
putBufs where in we join on this thread to complete. So, this results
in a deadlock scenario.

Fixed it by signalling the thread in putBuf() and setting buffers to
be allocated to 0 so that the thread exits as soon as possible.

Change-Id: Iecacdfc55ed85efbafa89ef97b0cdc7a740d8684
Signed-off-by: Ayush Dubey <ayushdubey70@gmail.com>